+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
   * new Backend WPTrackserver directly accessing its mysql database
   * new: Backend.clone()
   * new: Backend.needs_config
+  * new: Track.ids and gpxdo fix --id-from-backend
 
 1.3.2 release 2018-08-19
 ------------------------